Define a .item-flash css class that animates from a yellow to a transparent background in 2 seconds.
Define a .flash() method on TableRow, which adds this class, listens for the AnimationEnd event, and then removes the class.

Add a handler for the 'updated' signal on the TableRow's record, and call .flash() when it arrives.

Aside: .flash() should perhaps be implemented on UIWidget? (probably with a more non-clashing name..)